Refactored io converters
